projects/achievements-app/requirements.txt
# Основные зависимости
tkinter  # обычно входит в стандартную поставку Python
python-docx==0.8.11

# Для работы с данными
pytest==7.4.0  # для тестирования